* Source/DKArgument.[mh]: Implement marshalling of variant types. Also add the   ability to customize the unboxing of objects. * Source/DKEndpoint.m   Source/DKMethod.m   Source/DKPort.m:   Compile fixes for clang. * Source/DKOutgoingProxy.[mh]   Source/GNUmakefile:   Add stub implementation for outgoing proxies. * Source/DKProxy.m: Add method to distinguish outgoing and incoming proxies. * Source/Tests/TestDKProxy.m: Fix runtime.h include. * Source/Tests/TestDKArgument.m: Add test for custom unboxing methods. * Source/config.h.in   config.make.in   configure.ac   GNUmakefile.postamble:   Update configure script and paraphernalia to check for advanced runtime   features. * configure: Regenerate
Finish implementation of (un-)marshalling values between NSInvocation and D-Bus
messages. More flexible scheme for unboxing values. Various smaller fixes.


git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/dbuskit/trunk@30844 72102866-910b-0410-8b05-ffd578937521

+++ Source/DKPort.m
@@ -122,7 +122,7 @@ enum {
  * This is the main method used to dispatch stuff from the DO system to D-Bus.
  */
 - (BOOL)sendBeforeDate: (NSDate *)limitDate
-                 msgid: (NSUInteger)msgid
+                 msgid: (NSInteger)msgid
             components: (NSMutableArray *)components
 	          from: (NSPort *)recievePort
               reserved: (NSUInteger)reserverdHeaderSpace
@@ -215,7 +215,7 @@ enum {
 	         extra: (void*)extra
                forMode: (NSString*)mode
 {
-  NSLog(@"RunLoop events: Ignoring event of type %ld", type);
+  NSLog(@"RunLoop events: Ignoring event of type %llu", (unsigned long long)type);
 }
 
 /**
